+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-30T05:28:01.013Z",
+    "2026-04-30T08:46:57.988Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {